new EnergyResultFactory for creating a EnergyResult
